# chat-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A service for managing chats and messages",
        "title": "Chat Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/applications/{token}/chats": {
            "post": {
                "description": "Creates a new chat for an application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Create a new chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.CreateChatResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/applications/{token}/chats/{number}/messages": {
            "get": {
                "description": "Retrieves all messages from a specific chat",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Chat Number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/chat-service_internal_model.MessageResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new message in a specific chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Create a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Chat Number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message Content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.CreateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.CreateMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/applications/{token}/chats/{number}/messages/search": {
            "get": {
                "description": "Search for messages within a chat based on query text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Search messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Chat Number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search Query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/chat-service_internal_model.MessageResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Search query is required",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Chat not found",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "chat-service_internal_model.CreateChatResponse": {
            "type": "object",
            "properties": {
                "chat_number": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "chat-service_internal_model.CreateMessageRequest": {
            "type": "object",
            "required": [
                "body"
            ],
            "properties": {
                "body": {
                    "type": "string",
                    "example": "Welcome to instabug!!"
                }
            }
        },
        "chat-service_internal_model.CreateMessageResponse": {
            "type": "object",
            "properties": {
                "message_number": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "chat-service_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message here"
                }
            }
        },
        "chat-service_internal_model.MessageResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string",
                    "example": "Welcome to instabug!!"
                },
                "chat_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-11-19T20:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "number": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_handler.CreateMessageRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        }
    }
}